BLOscWithComplexSinusoid
This SuperCollider UGen is made for research purpose and is not an
efficient implementation. Unless you are interested in knowing Band
Limited Oscillator implementation with complex sinusoid, download
“BLOsc” and “BLOsc2,” but not this “BLOscWithComplexSinusoid.”

This is Band Limited Oscillator implementation with complex sinusoid
and without sinusoid wavetable. It works, but CPU usage is much heavier
than normal BLOsc. That is why it is not recommended for regular users.